# packages/shared-proto/openapi.yaml
openapi: 3.0.3
info:
  title: ACE API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /auth/register:
    post:
      deprecated: true
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      deprecated: true
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: ace_access
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          description: Plaintext password (sent over HTTPS in production)
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, instructor, admin]
        createdAt:
          type: string
          description: RFC3339 timestamp
...
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/User'